SQLite3 Multiple Ciphers 2.0.2 (based on SQLite 3.48.0)
* Based on SQLite version 3.48.0

Important
The SQL configuration functions can't be used any longer for configuring
the cipher scheme before issuing the PRAGMA key statement.

The reason is that starting with SQLite 3.48.0 executing a SELECT statement
will now always access the underlying database file. For encrypted databases
this will fail, because the cipher scheme is not yet activated.

Please use `PRAGMA statements or URI parameters for configuring the cipher scheme.