SQLite3 Multiple Ciphers 2.0.4 (based on SQLite 3.49.1)
* Based on SQLite version 3.49.1

Important:
The underlying SQLite version 3.49.1 fixes a bug discovered in the concat_ws()
SQL function. Malicious inputs to that function might lead to a memory error.
For that reason, you are encouraged to upgrade.
The problem has existed since the concat_ws() SQL function was first added
to SQLite with version 3.44.0 on 2023-11-01.